+++ lib/phoenix_kit_ecommerce/schemas/product.ex
@@ -33,6 +33,7 @@ defmodule PhoenixKitEcommerce.Product do
   """
 
   use Ecto.Schema
+  use PhoenixKit.SchemaPrefix
   import Ecto.Changeset
 
   @type t :: %__MODULE__{}